Migrate from CircleCI to Github Actions
The PR build (CI) is a reusable [required workflow](https://github.com/mattermost/actions-workflows/tree/main/.github/workflows/required_workflows/plugins). The CD
part is defined in the repository. In parallel we add the  to define the version of node we need to use to build.

Ticket: https://mattermost.atlassian.net/browse/CLD-4816